BotDiscord/requirements.txt
aiohttp~=3.11.18
beautifulsoup4~=4.13.4
python-dotenv~=1.1.0
discord.py==2.5.2